requirements.txt
numpy
pandas
scikit-learn
google-colab
